add 2 queries for JOIN

+++ migrations/assignments_submissions.sql
@@ -4,7 +4,7 @@ CREATE TABLE assignments (
   content TEXT,
   day INTEGER,
   chapter INTEGER,
-  duration INTEGER
+  duration FLOAT
 );
 
 
@@ -12,6 +12,6 @@ CREATE TABLE assignment_submissions (
   id SERIAL PRIMARY KEY NOT NULL,
   assignment_id INTEGER REFERENCES assignments(id) ON DELETE CASCADE,
   student_id INTEGER REFERENCES students(id) ON DELETE CASCADE,
-  duration INTEGER,
+  duration FLOAT,
   submission_date DATE
 );
